添加jwt认证

+++ configs/jwt.go
@@ -2,5 +2,5 @@ package configs
 
 type Jwt struct {
 	Secret string `yaml:"secret"`
-	JwtTtl int `yaml:"jwt_ttl"`
+	JwtTtl int64 `yaml:"jwt_ttl"`
 }
